Rename and add new Soroban types
